Reduce restrictions on pipeline model structure
Execute blocks in sequence, according to a topological sorted list.
This allows a pipeline model to have multiple starting blocks.
Additionally, blocks can have multiple outputs.

Blocks are still restricted to singular inputs as long as we have no
concept of how to assign inputs to parameters of the execute method.